Move native mode logic back
